requirements-dev.txt
-r requirements.txt
pre-commit==3.0.1
pytest-sugar==1.0.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
pytest-asyncio==0.23.6
